Refactor search results into distinct component; set up mobile view to use it
